save .env file - to complete - see description
By now, when the system simply creates a backup copy of the default .env file inside the root project dir. If the user install the package and have a .env.docker.backup file inside the dir, the file is copied as the default. This is good, but how to change settings in the .env? If you change it inside docker/.env, than the changes are not updated inside the .env.docker.backup file. I need to decide what to do: remove the .env.docker.backup from version control and backup it simply inside - for example - the data directory, adding a pre-update hook to have the most updated version? Or provide some convenient scripts to the user permitting to manage things automatically?